requirements.txt
Flask
Flask-Cors
Pillow
SPARQLWrapper
SQLAlchemy
Werkzeug
aiohttp
asyncio
backoff
beautifulsoup4
bigframes
chardet
charset-normalizer
colab-xterm
dask
dataclasses_json
datasets
deepeval
distro
docker
et-xmlfile
etils
faker
fastapi
filelock
firebase-admin
flatbuffers
functions-framework
functions_framework
gcloud
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-discoveryengine
google-cloud-documentai
google-cloud-enterpriseknowledgegraph
google-cloud-language
google-cloud-logging
google-cloud-pubsub
google-cloud-spanner
google-cloud-speech
google-cloud-storage
google-cloud-translate
google-cloud-videointelligence
google-cloud-vision
google-generativeai
googleapis-common-protos
gradio
gradio_webrtc
gspread
gunicorn
huggingface-hub
itables
itsdangerous
jinja2
joblib
jsonschema
jupyter-core
kfp
langchain
langchain-community
langchain-core
langchain-google-alloydb-pg
langchain-google-vertexai
langchain-text-splitters
langserve
librosa
llama-index
llama-index-embeddings-vertex
llama-index-llms-vertex
llama-index-storage-docstore-firestore
llama-index-utils-workflow
llama_deploy
lxml
markdown-it-py
mdurl
mesop
mpmath
mypy
nest-asyncio
numpy
oauth2client
opencv-python
opencv-python-headless
openpyxl
packaging
pandas
pillow
plotly
portpicker
praw
protobuf
psycopg2-binary
pyarrow
pydantic
pydantic-ai
pydub
pygments
pyparsing
pytest
pytest-asyncio
python-dateutil
python-docx
python-dotenv
pytz
pyvis
pyyaml
requests
rich
scikit-learn
scipy
seaborn
soundfile
soxr
streamlit
streamlit-agraph
streamlit-chat
streamlit-drawable-canvas
streamlit-extras
streamlit-navigation-bar
sympy
tblib
termcolor
text-generation
tiktoken
tokenizers
toml
toolz
torch
tqdm
traitlets
transformers
typing
typing-extensions
urllib3
vertexai
websockets
werkzeug
xgboost
xlsxwriter
xmltodict
yapf
